+++ src/Transformer/CustomTransformer/CustomTransformerInterface.php
@@ -10,11 +10,11 @@ namespace AutoMapper\Transformer\CustomTransformer;
 interface CustomTransformerInterface
 {
     /**
-     * @param mixed $input the source input (object or array) on which the custom transformation applies
+     * @param object|array<string, mixed> $source the source input on which the custom transformation applies
      *
      * Return value depends on the custom transformer's type:
-     * For CustomModelTransformerInterface, you should return a full target object,
-     * for CustomPropertyTransformerInterface, only the target property should be returned
+     *   - for CustomModelTransformerInterface, you should return a full target object,
+     *   - for CustomPropertyTransformerInterface, only the target property should be returned
      */
-    public function transform(mixed $input): mixed;
+    public function transform(object|array $source): mixed;
 }