+++ tests/AutoMapperBaseTest.php
@@ -6,6 +6,7 @@ namespace AutoMapper\Tests;
 
 use AutoMapper\AutoMapper;
 use AutoMapper\Extractor\ClassMethodToCallbackExtractor;
+use AutoMapper\Extractor\CustomTransformerExtractor;
 use AutoMapper\Generator\Generator;
 use AutoMapper\Loader\ClassLoaderInterface;
 use AutoMapper\Loader\FileLoader;
@@ -37,7 +38,7 @@ abstract class AutoMapperBaseTest extends TestCase
         $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $this->loader = new FileLoader(new Generator(
-            new ClassMethodToCallbackExtractor(),
+            new CustomTransformerExtractor(new ClassMethodToCallbackExtractor()),
             (new ParserFactory())->create(ParserFactory::PREFER_PHP7),
             new ClassDiscriminatorFromClassMetadata($classMetadataFactory),
             $allowReadOnlyTargetToPopulate