+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [GH#25](https://github.com/jolicode/automapper/pull/25) Pass full input object to property custom transformers
 - [GH#10](https://github.com/jolicode/automapper/pull/10) Introduce custom transformers
 
 ## [8.1.0] - 2023-12-14